Refactor database handling and restructure API routes; add authentication and book management routers
