+++ configs/Database.py
@@ -1,6 +1,9 @@
+from typing import Annotated
+from sqlalchemy.orm import Session
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 from sqlalchemy.ext.declarative import declarative_base
+from fastapi import Depends
 
 SQLALCHEMY_DATABASE_URL = 'sqlite:///./metadata/book.db'
 
@@ -9,3 +12,15 @@ engine = create_engine(SQLALCHEMY_DATABASE_URL, connect_args={'check_same_thread
 SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
 
 Base = declarative_base()
+
+def get_db():
+    """
+    Start DataBase
+    """
+    db = SessionLocal()
+    try:
+        yield db
+    finally:
+        db.close()
+
+db_dependency = Annotated[Session, Depends(get_db)]
