+++ routers/v1/AuthRouter.py
@@ -1,0 +1,10 @@
+from fastapi import APIRouter
+
+AuthRouter = APIRouter()
+
+@AuthRouter.get("/user/")
+async def get_user():
+    """
+    Autenticate User
+    """
+    return {'user': 'autenticated'}
